+++ internal/cli/atlas/dbusers/certs/list.go
@@ -12,41 +12,34 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package atlas
+package certs
 
 import (
 	"github.com/mongodb/mongocli/internal/cli"
 	"github.com/mongodb/mongocli/internal/config"
-	"github.com/mongodb/mongocli/internal/convert"
 	"github.com/mongodb/mongocli/internal/description"
 	"github.com/mongodb/mongocli/internal/flag"
 	"github.com/mongodb/mongocli/internal/json"
 	"github.com/mongodb/mongocli/internal/store"
 	"github.com/mongodb/mongocli/internal/usage"
 	"github.com/spf13/cobra"
-	atlas "go.mongodb.org/atlas/mongodbatlas"
 )
 
-type DBUsersUpdateOpts struct {
+type ListOpts struct {
 	cli.GlobalOpts
+	cli.ListOpts
+	store    store.UserCertificateDescriber
 	username string
-	password string
-	roles    []string
-	store    store.DatabaseUserUpdater
 }
 
-func (opts *DBUsersUpdateOpts) initStore() error {
+func (opts *ListOpts) initStore() error {
 	var err error
 	opts.store, err = store.New(config.Default())
 	return err
 }
 
-func (opts *DBUsersUpdateOpts) Run() error {
-	current := new(atlas.DatabaseUser)
-
-	opts.update(current)
-
-	result, err := opts.store.UpdateDatabaseUser(current)
+func (opts *ListOpts) Run() error {
+	result, err := opts.store.GetUserCertificates(opts.ConfigProjectID(), opts.username)
 
 	if err != nil {
 		return err
@@ -55,38 +48,22 @@ func (opts *DBUsersUpdateOpts) Run() error {
 	return json.PrettyPrint(result)
 }
 
-func (opts *DBUsersUpdateOpts) update(out *atlas.DatabaseUser) {
-	out.GroupID = opts.ConfigProjectID()
-	out.Username = opts.username
-	if opts.password != "" {
-		out.Password = opts.password
-	}
-
-	out.Roles = convert.BuildAtlasRoles(opts.roles)
-}
-
-// mongocli atlas dbuser(s) update <username> [--password password] [--role roleName@dbName] [--projectId projectId]
-func DBUsersUpdateBuilder() *cobra.Command {
-	opts := &DBUsersUpdateOpts{}
+// mongocli atlas dbuser(s) certs list|ls <username> [--projectId projectId]
+func ListBuilder() *cobra.Command {
+	opts := &ListOpts{}
 	cmd := &cobra.Command{
-		Use:   "update <username>",
-		Short: description.UpdateDBUser,
-		Example: `
-  Update roles for a user
-  $ mongocli atlas dbuser update <username> --role readWriteAnyDatabase --projectId <projectId>`,
-		Args: cobra.ExactArgs(1),
+		Use:   "list <username>",
+		Short: description.ListDbUersCerts,
+		Args:  cobra.ExactArgs(1),
 		PreRunE: func(cmd *cobra.Command, args []string) error {
+			opts.username = args[0]
+
 			return opts.PreRunE(opts.initStore)
 		},
 		RunE: func(cmd *cobra.Command, args []string) error {
-			opts.username = args[0]
 			return opts.Run()
 		},
 	}
-
-	cmd.Flags().StringVarP(&opts.username, flag.Username, flag.UsernameShort, "", usage.Username)
-	cmd.Flags().StringVarP(&opts.password, flag.Password, flag.PasswordShort, "", usage.Password)
-	cmd.Flags().StringSliceVar(&opts.roles, flag.Role, []string{}, usage.Roles)
 
 	cmd.Flags().StringVar(&opts.ProjectID, flag.ProjectID, "", usage.ProjectID)
