+++ internal/cli/atlas/dbusers/dbusers_create.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package atlas
+package dbusers
 
 import (
 	"errors"
