+++ internal/cli/atlas/dbusers/dbusers_update_test.go
@@ -14,7 +14,7 @@
 
 // +build unit
 
-package atlas
+package dbusers
 
 import (
 	"testing"
