+++ internal/description/description.go
@@ -69,6 +69,8 @@ const (
 	CreateDBUser             = "Create a database user for a project."
 	DeleteDBUser             = "Delete a database user for a project."
 	ListDBUsers              = "List Atlas database users for a project."
+	DbUserCerts              = "Manage Atlas x509 certificates for your database users."
+	ListDbUersCerts          = "Get a list of all Atlas-managed, unexpired certificates for a database user."
 	ListEvents               = "List events for an organization or project."
 	UpdateDBUser             = "Update a MongoDB dbuser in Atlas."
 	ProcessMeasurements      = "Get measurements for a given host."
